pyproject.toml
[tool.poetry]
name = "exchange"
version = "0.1.0"
description = "A dummy asset exchange used in my talks as a livecoding/livedemo basis"
authors = ["Johnny Rocketfingers <developer94@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.6"
djangorestframework = "^3.12.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-vcr = "^1.0.2"
pytest-django = "^4.1.0"
flake8 = "^3.8.4"
black = "^20.8b1"
django-stubs = "^1.7.0"
model-bakery = "^1.2.1"
djangorestframework-stubs = "^1.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
